postremove fixes

+++ pkgsrc/postinstall.sh
@@ -2,7 +2,7 @@
 
 set -xeu
 
-mv /var/run/zerotier-one.service /lib/systemd/system
+test -f /var/run/zerotier-one.service && mv /var/run/zerotier-one.service /lib/systemd/system
 systemctl daemon-reload
 systemctl enable zerotier-systemd-manager.timer
 systemctl start zerotier-systemd-manager.timer
